Create variable unique_domain_validation_options
The domain_validation_options is a list of maps of the validation
options needed for all the domains. The wildcard domains use the same
validation options as the APEX domain though.

The new unique_domain_validation_options variable is created from the
domain_validation_options list, without the wildcard validation options,
so that we avoid duplicates.